+++ R/context.R
@@ -181,6 +181,9 @@ reply <- function(context,
 #' @param cv (optional) a 'conditionVariable' to signal when the async receive
 #'   is complete, or NULL. If any other value is supplied, this will cause the
 #'   pipe connection to be dropped when the async receive is complete.
+#' @param msgid (optional) integer message ID to send a special payload to the
+#'   context upon timeout (asynchronously) consisting of an integer zero,
+#'   followed by the value of `msgid` supplied.
 #'
 #' @return A 'recvAio' (object of class 'mirai' and 'recvAio') (invisibly).
 #'
@@ -227,5 +230,6 @@ request <- function(context,
                     recv_mode = c("serial", "character", "complex", "double",
                                   "integer", "logical", "numeric", "raw", "string"),
                     timeout = NULL,
-                    cv = NULL)
-  data <- .Call(rnng_request, context, data, send_mode, recv_mode, timeout, cv, environment())
+                    cv = NULL,
+                    msgid = NULL)
+  data <- .Call(rnng_request, context, data, send_mode, recv_mode, timeout, cv, msgid, environment())