+++ src/sync.c
@@ -16,12 +16,25 @@ static void sendaio_complete(void *arg) {
 static void request_complete(void *arg) {
 
   nano_aio *raio = (nano_aio *) arg;
+  nano_saio *saio = (nano_saio *) raio->cb;
   int res = nng_aio_result(raio->aio);
   if (res == 0) {
     nng_msg *msg = nng_aio_get_msg(raio->aio);
     raio->data = msg;
     nng_pipe p = nng_msg_get_pipe(msg);
     res = - (int) p.id;
+  } else if (res == 5) {
+    const int id = saio->msgid;
+    if (id) {
+      nng_msg *msg;
+      if (nng_msg_alloc(&msg, 0) == 0) {
+        if (nng_msg_append_u32(msg, 0) ||
+            nng_msg_append(msg, &id, sizeof(id)) ||
+            nng_ctx_sendmsg(*saio->ctx, msg, 0)) {
+          nng_msg_free(msg);
+        }
+      }
+    }
   }
 
   if (raio->next != NULL) {
@@ -38,7 +51,6 @@ static void request_complete(void *arg) {
     raio->result = res;
   }
 
-  nano_saio *saio = (nano_saio *) raio->cb;
   if (saio->cb != NULL)
     later2(raio_invoke_cb, saio->cb);
 
@@ -374,13 +386,14 @@ SEXP rnng_cv_signal(SEXP cvar) {
 
 // request ---------------------------------------------------------------------
 
-SEXP rnng_request(SEXP con, SEXP data, SEXP sendmode, SEXP recvmode, SEXP timeout, SEXP cvar, SEXP clo) {
+SEXP rnng_request(SEXP con, SEXP data, SEXP sendmode, SEXP recvmode, SEXP timeout, SEXP cvar, SEXP msgid, SEXP clo) {
 
   if (NANO_PTR_CHECK(con, nano_ContextSymbol))
     Rf_error("'con' is not a valid Context");
 
   const nng_duration dur = timeout == R_NilValue ? NNG_DURATION_DEFAULT : (nng_duration) nano_integer(timeout);
   const uint8_t mod = (uint8_t) nano_matcharg(recvmode);
+  const int id = msgid != R_NilValue ? NANO_INTEGER(msgid) : 0;
   int signal, drop;
   if (cvar == R_NilValue) {
     signal = 0;
@@ -402,6 +415,8 @@ SEXP rnng_request(SEXP con, SEXP data, SEXP sendmode, SEXP recvmode, SEXP timeou
   nano_encodes(sendmode) == 2 ? nano_encode(&buf, data) : nano_serialize(&buf, data, NANO_PROT(con));
   saio = R_Calloc(1, nano_saio);
   saio->cb = NULL;
+  saio->ctx = ctx;
+  saio->msgid = id;
 
   if ((xc = nng_msg_alloc(&msg, 0)))
     goto exitlevel1;
@@ -434,6 +449,7 @@ SEXP rnng_request(SEXP con, SEXP data, SEXP sendmode, SEXP recvmode, SEXP timeou
   PROTECT(env = R_NewEnv(R_NilValue, 0, 0));
   Rf_classgets(env, nano_reqAio);
   Rf_defineVar(nano_AioSymbol, aio, env);
+  Rf_setAttrib(env, nano_MsgidSymbol, msgid);
 
   PROTECT(fun = R_mkClosure(R_NilValue, nano_aioFuncMsg, clo));
   R_MakeActiveBinding(nano_DataSymbol, fun, env);